requirements.txt
asgiref==3.3.1
ccy==1.2.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
djangorestframework==3.12.2
forex-python==1.5
gzinfo==1.0.2
idna==2.10
numpy==1.20.1
pandas==1.2.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
